# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: AnimusForge API
  description: |
    ## Adaptive AI Persona Framework with Ethical Governance

    AnimusForge is an enterprise-grade framework for creating, managing, and evolving
    AI personas with built-in ethical governance, kill-switch mechanisms, and
    comprehensive memory systems.

    ### Key Features

    - **Persona Management**: Create, update, activate, and evolve AI personas with
      configurable traits, cognitive patterns, and behavioral constraints.
    - **LLM Gateway**: Multi-provider LLM access with automatic failover, load balancing,
      and cost optimization across OpenAI, Anthropic, DeepSeek, and GLM.
    - **Kill-Switch**: Emergency stop and recovery mechanisms for immediate persona
      deactivation with audit trail and recovery workflows.
    - **Memory System**: Vector-based semantic memory and graph-based knowledge storage
      for persistent persona context and learning.
    - **Ethics Engine**: Real-time content evaluation, audit logging, and compliance
      reporting across multiple ethical categories.
    - **System Monitoring**: Health checks, Prometheus metrics, and configuration
      management endpoints.

    ### Authentication

    JWT authentication is supported. Include `Authorization: Bearer <token>` header
    for authenticated requests. Some endpoints may require elevated permissions.

    ### Rate Limiting

    All endpoints are rate-limited. Check response headers:
...
  version: 1.0.0
  contact:
...
servers:
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT authentication token
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for service-to-service authentication
  schemas:
...
    UUID:
      type: string
      format: uuid
      example: "550e8400-e29b-41d4-a716-446655440000"
    Timestamp:
      type: string
      format: date-time
      example: "2024-03-06T12:00:00Z"
...
    MemoryType:
      type: string
      enum:
        - episodic
        - semantic
        - procedural
        - working
        - long_term
    VectorStoreRequest:
      type: object
      required:
        - persona_id
        - content
      properties:
        persona_id:
          $ref: "#/components/schemas/UUID"
        content:
          type: string
          minLength: 1
        memory_type:
          $ref: "#/components/schemas/MemoryType"
        embedding:
          type: array
          items:
            type: number
            format: float
        metadata:
          type: object
        tags:
          type: array
          items:
            type: string
        importance:
          type: number
          format: float
          minimum: 0
          maximum: 1
          default: 0.5
        ttl_seconds:
          type: integer
    VectorSearchRequest:
      type: object
      required:
        - persona_id
        - query
      properties:
        persona_id:
          $ref: "#/components/schemas/UUID"
        query:
          type: string
          minLength: 1
        memory_types:
          type: array
          items:
            $ref: "#/components/schemas/MemoryType"
        tags:
          type: array
          items:
            type: string
        top_k:
          type: integer
          default: 10
          minimum: 1
          maximum: 100
        min_score:
          type: number
          format: float
          default: 0
    VectorSearchResult:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/UUID"
        score:
          type: number
          format: float
        content:
          type: string
        memory_type:
          $ref: "#/components/schemas/MemoryType"
        created_at:
          $ref: "#/components/schemas/Timestamp"

    # ==================== Ethics Schemas ====================
...
paths:
...
  /api/v1/memory/vector/store:
    post:
      tags:
        - Memory
      summary: Store vector
      description: Store a vector embedding with associated content and metadata.
      operationId: storeVector
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VectorStoreRequest"
      responses:
        "201":
          description: Vector stored successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    $ref: "#/components/schemas/UUID"
                  persona_id:
                    $ref: "#/components/schemas/UUID"
                  stored:
                    type: boolean
                  dimension:
                    type: integer
                  created_at:
                    $ref: "#/components/schemas/Timestamp"
  /api/v1/memory/vector/search:
    post:
      tags:
        - Memory
      summary: Search vectors
      description: Search for similar vectors using semantic similarity.
      operationId: searchVectors
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VectorSearchRequest"
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                type: object
                properties:
                  results:
                    type: array
                    items:
                      $ref: "#/components/schemas/VectorSearchResult"
                  total:
                    type: integer
                  query_time_ms:
                    type: number
  /api/v1/memory/graph/node:
    post:
      tags:
        - Memory
      summary: Create graph node
      description: Create a new node in the knowledge graph.
      operationId: createGraphNode
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - persona_id
                - node_type
                - name
              properties:
                persona_id:
                  $ref: "#/components/schemas/UUID"
                node_type:
                  type: string
                  enum: [entity, concept, event, persona, interaction, knowledge, skill, context]
                name:
                  type: string
                properties:
                  type: object
                content:
                  type: string
      responses:
        "201":
          description: Node created successfully
  /api/v1/memory/graph/node/{node_id}:
    get:
      tags:
        - Memory
      summary: Get graph node
      description: Retrieve a specific node from the knowledge graph.
      operationId: getGraphNode
      parameters:
        - name: node_id
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/UUID"
      responses:
        "200":
          description: Node details
        "404":
          description: Node not found
  /api/v1/memory/graph/relationship:
    post:
      tags:
        - Memory
      summary: Create graph relationship
      description: Create a relationship between two nodes in the knowledge graph.
      operationId: createGraphRelationship
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - persona_id
                - source_id
                - target_id
                - relationship_type
              properties:
                persona_id:
                  $ref: "#/components/schemas/UUID"
                source_id:
                  $ref: "#/components/schemas/UUID"
                target_id:
                  $ref: "#/components/schemas/UUID"
                relationship_type:
                  type: string
                  enum: [relates_to, derived_from, depends_on, caused_by, follows, precedes, contains, references]
                weight:
                  type: number
                  default: 1.0
                bidirectional:
                  type: boolean
                  default: false
      responses:
        "201":
          description: Relationship created successfully
        "404":
          description: Source or target node not found
  /api/v1/memory/graph/query:
    get:
      tags:
        - Memory
      summary: Query graph
      description: Execute a query against the knowledge graph.
      operationId: queryGraph
      parameters:
        - name: persona_id
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/UUID"
        - name: query_type
          in: query
          schema:
            type: string
            enum: [cypher, traverse, pattern, neighbors]
            default: traverse
        - name: query
          in: query
          required: true
          schema:
            type: string
        - name: max_depth
          in: query
          schema:
            type: integer
            default: 3
            minimum: 1
            maximum: 10
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
      responses:
        "200":
          description: Query results

  # ==================== Ethics ====================
...